t2130: decompose doc/agent indexing python cluster into shared modules
Extract shared logic from 5 high-complexity Python files into 4 focused
library modules under .agents/scripts/lib/:

- agent_config.py: shared agent constants, discovery loop, subagent validation
  (deduplicated from agent-discovery.py and opencode-agent-discovery.py)
- mcp_config.py: MCP registration and loading policy helpers
  (extracted from both discovery scripts)
- doc_linking.py: Document class and relationship builders
  (extracted from cross-document-linking.py)
- entity_cleaners.py: entity cleaning, LLM response parsing, regex extraction
  (extracted from entity-extraction.py)

File complexity reductions:
- cross-document-linking.py: 524 → 328 lines (-37%)
- opencode-agent-discovery.py: 693 → 137 lines (-80%)
- entity-extraction.py: 524 → 329 lines (-37%)
- agent-discovery.py: 540 → 271 lines (-50%)
- add-related-docs.py: split build_navigation_section into 3 focused helpers

Resolves #19226
